Update patch set 4
Patch Set 4:

I wonder if this is a full solution.

zuul doesn't differentiate projects based on their connections. It uses the source of the pipeline when setting it up, but it'll assume a project is the same project regardless of the pipeline. This may not be something worth supporting, but I haven't looked at what would happen in this scenario.

Patch-set: 4
Label: Verified=0