Update patch set 4
Patch Set 4:

(1 comment)

> So why not also look for ZUUL_NEWREV?

Just to check that I'm understanding you here, do you mean it should look for ZUUL_NEWREV env-var and try to check everything out as a the indicated_revision?


While both are benign, doing so feels a bit dirty, as if ZUUL_NEWREV is set and if we check out a lot of repos, we'll end up throwing a lot of error messages, and waste time attempting to checkout something that (hopefully) does not exist.

However, pulling from the ZUUL_NEWREV variable seems to fit in with the current workflow of zuul-cloner, so I'll get that added.

Patch-set: 4