Update patch set 3
Patch Set 3: Code-Review-1

I believe this is already supported.. Maybe I'm missing something so just -1 for now.

http://git.openstack.org/cgit/openstack-infra/zuul/tree/zuul/lib/swift.py#n41

If you set X-Storage-Url and Send-Temp-Url-Key is set to False, no connection to swift should be made.

We could perhaps default Send-Temp-Url-Key to False when X-Account-Meta-Temp-Url-Key is supplied. Currently it defaults to True which will send the temp-url key regardless if it is defined in the config or generated by itself. (Which is basically to ensure the credentials are set correctly).

This is documented here: http://docs.openstack.org/infra/zuul/zuul.html#swift

Patch-set: 3
Label: Code-Review=-1