Don't launch jobs in threads
If a job fails in a thread the exception won't be caught by the above
run method. The server itself is already ran in a thread so unless we
want multiple jobs to be run by the server we probably don't want to
thread the job itself.

If we do want multiple jobs to be ran in parallel I think the solution
should be to just run multiple zuul worker programs.

Change-Id: I42dabc034790b8448d1d5cf1e9f57987a445ffc8