Use the seeded repos when launching a job
The launcher would check out each repo but then prepare the project
in a jobdir git path which would reclone the repository from the
item url (and hence not use any local cache).

Instead use the local git copy from the launcher for each job build.
We need to reset the repo to master (or possible a branch, see note)
at the beginning of each job in case it was left in an undesirable
state from a previous run.

We reset because a project may use ansible configuration from
another repo that is not a dependant change (and therefore not in
the job items).

We may need to consider how to clean up all of the changes/refs
that a worker will eventually collect.

Change-Id: I1b661b886a367116d114a23be72511257a08caeb