Update patch set 4
Patch Set 4:

(1 comment)

I would have to double check the code, but I think in the case of using ZUUL_NEWREV it'll only be set for the primary change (and no dependant ones). In fact, I'm not even sure the post pipeline can have dependant changes since all of its dependencies will have been merged.

In other words we don't need to check for the ref from every project attempting to be cloned if we look at ZUUL_CHANGE.

Patch-set: 4