Update patch set 9
Patch Set 9: Code-Review-1

Thanks, I agree this is a problem.  :)

I think it's important for developers to still be able to run one test at a time with full debug logging.  That is how I test changes I am making to Zuul (and I hope others do too) -- it's about the only way that a developer can see what Zuul is actually doing (and what the logs will look like) before a change lands.

What I think would be best would be to discard the log entries except on tests that fail.  I don't know if that's possible or easy.

If it's not, then defaulting to INFO is okay, but we should allow developers to override locally with an environment variable.

Patch-set: 9
Label: Code-Review=-1