Update patch set 8
Patch Set 8:

(1 comment)

> (1 comment)
 > 
 > -1 for in-line comment about testing.
 
I will do this.

 > But also, if zuul-cloner is already aware of the zuul pipeline and
 > the new revision from environment variables, why force the jobs to
 > add logic to pass in the correct git sha? I have a change up that
 > works in both post and release pipelines at https://review.openstack.org/271585

As mentioned in that review, I do not think that it a solution to the post jobs.  We tried that method in gate by setting ZUUL_BRANCH = ZUUL_REF in project-config.  The results were different/incorrect compared to what gerrit-git-prep produced and we had to switch back to the old post jobs.

 > I think this feature to specify explicit shas is useful, but not
 > necessarily needed to get zuul-cloner working in post pipelines.

You don't need to explicitly set the sha, but it is an option.  If the 'ZUUL_NEWREV' variable is defined, we will grab via the SHA.

Patch-set: 8