Clean up results returned from ansible-launcher
The gearman launcher expects the job to return a dictionary
which may contain extra information. Provide this in case
it becomes useful. Also makes the response more predictable.

The sendCompleteEvent would take serialised json and place
it into a dict before double-serialising it all to send a
zmq message. Instead just pass the actual result.

If the runJob fails during setup no result is returned.
This is confusing the scheduler so we should try and
return UNSTABLE. We could also consider using
sendWorkException packet but this is currently
compatible with looking like Jenkins.

Change-Id: I897a7e9476122c31870a68f0e4ae1f68944d5248