Let paramfunc specify job name dynamically
To reduce the number of Jenkins jobs, several of them could be run under
the same name if the parameters differ. To solve this, the python
parameter function that exists in Zuul layout.yaml replaces the job name
in the configuration with the job name that should be started on
Jenkins. The Zuul report still reports the result under the original
name specified in layout.yaml.

Signed-off-by: Fredrik Medley <fredrik.medley@autoliv.com>
Signed-off-by: Fredrik Medley <fredrik.medley@gmail.com>
Change-Id: Ibc973a944b4dbcf1e8432681196a5e21d0a0e352