Update patch set 2
Patch Set 2:

Just to reiterate what I said above - without this patch, in a case where I have private Gerrit repos, I'm getting this error:

  2016-05-18 09:28:25,976 ERROR zuul.Gerrit: Exception looking for ref refs/heads/master
  Traceback (most recent call last):
    File "/usr/lib/python2.7/site-packages/zuul/trigger/gerrit.py", line 211, in getRefSha
      refs = self._getInfoRefs(project)
    File "/usr/lib/python2.7/site-packages/zuul/trigger/gerrit.py", line 176, in _getInfoRefs
      raise Exception("Gerrit repository does not support "
  Exception: Gerrit repository does not support git-upload-pack

With this patch, the error is gone. That's on the 2.1.0 version of Zuul, not on Master, but when applying the patch, I refactored it correctly, to fit 2.1.0 code.

Patch-set: 2