Update patch set 9
Patch Set 9: Code-Review-1

I understand why you implemented it as "project_revisions[project] = newrev" (in order to reduce spurious error output when checking out multiple projects).  But I don't see a need to actually expose the --project-revision option.  I would like to keep this as simple as possible (believe it or not, it is currently as simple as possible to handle all the known use cases), and so don't want to add options that might make job configurations more confusing later.

Is there a strict need for that added command line option (i.e., a use case that can not be served without it), or can we omit it for now?

Patch-set: 9
Label: Code-Review=-1