Update patch set 12
Patch Set 12: Code-Review-1

(3 comments)

Cool.  I have a minor code reorg change suggested inline, along with a potential behavior change we should consider (definitely want others to weigh in on fetch comment).  I also added a change that depends on this one that does as thorough of a simulation of how we will actually end up using this (including with cached repos) as possible.  It presupposes that you accept the reorg change, so will fail right now, but otherwise passes when modified to account for that (yay!).

Patch-set: 12
Label: Code-Review=-1