Update patch set 12
Patch Set 12:

(3 comments)

> (3 comments)
 > 
 > Cool.  I have a minor code reorg change suggested inline, along
 > with a potential behavior change we should consider (definitely
 > want others to weigh in on fetch comment).

I've added the fetch, however I think moving the dictionary into lib/cloner is not the right thing to do.  It moves shell logic into the lib, and also requires passing ZUUL_PROJECT in as well.  See comments in cmd.cloner.

 > I also added a change
 > that depends on this one that does as thorough of a simulation of
 > how we will actually end up using this (including with cached
 > repos) as possible.

I'm assuming this is in relation to the potential for a fetch failure.  It seems to pass regardless of a fetch or not, when made to run against the current state of the patch-set.

Patch-set: 12
Label: Verified=0