Call :func:`AnsibleModule.fail_json` on failure
When the Zuul task is handled by the rescue(error) block, make sure
the task is marked as a failure via calling the `.fail_json` method
instead of the `.exit_json` method.

This adds an optional "failure" param to the `zuul_log` module that
gates the failure status vs standard exit status.

Change-Id: I07d2f06aa74cfe9dd0f7a1c3a64d4e51290bd570
Partial-Story: #2000619