Update patch set 2
Patch Set 2:

recheck

I wish I knew what was causing the error too.  I am finding the testtools timeout fixtures to be somewhat difficult to use.  The soft timeout does not cause our tests to stop and so defeats the point.  The hard timeout (which we use instead) causes us to end up with no logs of the failure.  So I have no way of knowing what is causing this test to be flaky.  It did not use to be, so I feel confident it is exposing some fault that we have introduced either in the test or in the server.  I have run this test over a weekend locally with no failures.  If I had a log from a failure, I am certain that I could correct the problem.

It would be very useful to figure out how we could use the soft timeout fixture, or, at this point, if there is another test running system that can handle timeouts and get us logs in that case, I am desperate enough to switch to it.

Patch-set: 2