Python 3 Fixes: Do not contest locks
In PythonGit the config writer was getting bound up in locking
dealing writing the configuration out. This moves to the context
manager that ensures write() is called and explicit lock release
is happening. The difference is in how py27 and py3 handle the
accquire lock and it appears py27 was not really enfocing
lock accquisition in how the zuul tests were being run.

Change-Id: I843f3aa60650f83188ed1e78730d19e4c2c888e0