Update patch set 4
Patch Set 4: Code-Review-1

I also think that it would be good to avoid having the configuration hard-coded into this.  I think we could take the approach we have with "copy-console: true" and add a "ansible-configuration" publisher with "copy-ansible: true" to our JJB config, and then in ansiblelauncher, look for copy-ansible and switch on that logic like we do for console logs.

That would mean changes to all of the jjb jobs, so if you didn't want to wait for that, as an interim step, we could add a temporary function to ansiblelauncher that automatically inserted that config for every job if it didn't have it already, and then we can remove that once the jjb config has landed.  The interim function will be nicely compartmentalized.

How does that sound?

Patch-set: 4
Label: Code-Review=-1