Update patch set 22
Patch Set 22:

> Instead of passing the webapp to connections, why not have the
 > webapp call a hook on the connections registered with the
 > scheduler?

It seemed natural to let the connections (un)register the hooks whenever they need to, eg. in the onLoad and onStop hooks.

If I invent another callbacks on connections for (un)registering the http handlers, what should be the order of the callbacks? Call connection.registerHttpHooks(webapp) before onLoad and connection.unregisterHttpHooks(webapp) after onStop? Or the other way?

Also instead of storing the WebApp instance on connection, I'd have to store it on the scheduler, because of the stopConnections call, which should result in unregistering the http handlers from connections.

Do you have any particular benefits of the suggested solution in mind?

Patch-set: 22