Update patch set 2
Patch Set 2:

[This comment is a comment on a comment, not on this patch.]

For the record, I am *strongly* against adding an extra log level -- I think info should be enough for operators to understand what's going on and debug should be enough for a core developer to figure out what went wrong post-facto.  I don't think we need a level that is "more information than could be useful for a post-mortem analysis".

Patch-set: 2