DNM: TESTING FIX: Capture Logs when Timeout is hit
Capture the logs when timeout is hit on a given test. This utilizes
a direct write to sys.stderr. This means that the standard capture
mechanisms are not there the log streams only to the console not
to the normal testrepository output.

Eventually "gentle" should be used instead of the standard
SIGALRM action, but this guarentees continued proper exit
on tests in the case the threads do not close down
elegantly on an exception.

Change-Id: I4ea408c5d59aa0361e3ca1ea8177269ab4820f02