Update patch set 1
Patch Set 1: Code-Review-1

That's not what this argument does at all.  The argument is "-e timeout=x" which sets the ansible variable "timeout" to "x" which is used by the zuul timeout callback plugin to keep a running timeout of the whole job (so that if we have 2 shell steps that run 30 minutes each, but a timeout for the whole job of 45 minutes, we still time out halfway through the second shell step).

Patch-set: 1
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1