Update patch set 4
Patch Set 4:

I spoke with lifeless and I have further things to improve upon here. But I'd like to land this as is first so we are seeing the errors directly. It will lays the foundation for the longer term solution. Including detecting threads that aren't in a daemon mode.

Patch-set: 4
Label: Verified=0