Add human-log filter plugin
This callback plugin filters the output to make it more human
readable. [1] is an example of a rather un-readable log I was
debugging for Ia034c5176937f579e0f19504c6f0ef9c118869b8

One thing to be aware is that this duplicates the log-output, dmsimard
gave an example at [2].  So "nicer" output is kind of subjective in
this regard.  Personally, I would have found it nicer to have the
"human readable" form quickly available.

The root of the discusion is [3]; this imports the latest plugin which
I have only made minor changes to for pep8 conformance.

[1] http://logs.openstack.org/89/330789/1/check/gate-tempest-dsvm-ironic-agent_ssh-coreos-src/7306594/_zuul_ansible/logs/ansible.log
[2] https://etherpad.openstack.org/p/TvALrzw4ww
[3] http://blog.cliffano.com/2014/04/06/human-readable-ansible-playbook-log-output-using-callback-plugin/

Change-Id: Icdebe1fdba1c30908e7ed475dff0c85d3b307eda