WIP: Integrate ARA reporting inside Zuul Launcher
This is mostly to see and discuss what work would be involved in
integrating ARA into Zuul.

ARA is a new project that's aimed at providing an interface to
browse Ansible Playbook runs and it looks like this [1].

There is value in providing an easy to parse interface when
Ansible output is either too verbose or not verbose enough.

ARA is an alternative or a supplement to human_log [2].

[1]: http://logs.openstack.org/34/330834/3/check/gate-ara-integration-centos-7/0df690f/logs/build/
[2]: https://review.openstack.org/#/c/330856/

Change-Id: Ibd52eef0cb84cbe8dbdf0374d424dfdc289f7928