Update patch set 14
Patch Set 14:

I'm ok with doing the subclass work here. I was planning on providing a cleaner interface directly in gear as a supplement to the current interfaces and using that in the ZuulV3 work that will spin off this. Since this work has to be replicated anyway in V3 branches (I assume that is still the plan) it might make sense to do that there. I'll mull it over tonight/this weekend.

Patch-set: 14
Reviewer: Gerrit User 2903 <2903@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0