Move the command processing into the launcher cmd
The launcher is who reads in zuul.conf and needs to in order to set
some basics such as pidfile etc.

Previously the reconfigure command would just reload the JJB config
for zuul launcher. This is slightly misleading as you might expect
a reconfigure to re-read the config (similar to how a SIGHUP does
on the scheduler). Moving the commands into the launcher cmd allows
us to reload the config.

This allows us to reconfigure things such as the zuul sites.

Change-Id: I57a36a97bfccd15c6552355bceb5d9e96909b027