Update patch set 2
Patch Set 2:

My main reason for trying it was to reduce the potential issue we've seen around SSH connection and file not found errors.

While async doesn't appear to support it ( I am going to confirm this now), I do think we should have the discussion around SSH connections.

But I'll defer to the group if we think it is worth the effort to run this in zuulv2.5 or not.

Patch-set: 2
Reviewer: Gerrit User 4162 <4162@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0