Update patch set 2
Patch Set 2: Code-Review+2

My bad. This used to not work in the old version of the timeout plugin where we were reading the extra_vars in the plugin to get and calculate the timeout. However, the plugin now only does elapsed_time and reports it with the calculation in the jinja - so this totally works now.

Patch-set: 2
Reviewer: Gerrit User 2 <2@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2