Make job registration with labels optional
Label registrations act as a multiplier against the number of jobs we
need to register. In many cases they are not required so make them
optional in an effort to reduce the total number of registrations
necessary.

This is implemented by adding a zuul launcher config flag that toggles
jobs with label registration. It is enabled by default for backward
compat.

Change-Id: If778312c161c72bb8a73e8638df2f09e89408e46