Update patch set 1
Patch Set 1:

Sachi, I'm not certain I follow -- the logic this tests is unchanged, and has been correct for some time.  This test did not match the API at the time it was written, but if either the API or the test were changed to be compatible, then it was expected to pass (and did in my experience).  Now there is a new revision of the parent change with an API that is nearly, but not quite, what this test anticipated.  With the small modification in PS2, it passes for me locally.

Patch-set: 1
Label: Verified=0