Update patch set 11
Patch Set 11: Code-Review-1

Just wondering if you had any thoughts on my comment on ps4?

Quoted:

I wonder if this is a full solution.
zuul doesn't differentiate projects based on their connections. It uses the source of the pipeline when setting it up, but it'll assume a project is the same project regardless of the pipeline. This may not be something worth supporting, but I haven't looked at what would happen in this scenario.

Patch-set: 11
Reviewer: Gerrit User 7069 <7069@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1