Update patch set 1
Patch Set 1:

Context:
We're using Zuul with zuul_swift_upload.

We'd like to have child jobs consume artifacts from the parent job. 

However, since the LOGPATH uses ZUUL_UUID, we have no way of retrieving the parent job ZUUL_UUID from a child job as that parameter is unique per-job.

So we'd like to be able to override LOG_PATH [1] to use the uuid in ZUUL_REF instead of ZUUL_UUID since ZUUL_REF is unique per build set (so the same for parent and children jobs).
We tried to override it through openstack_functions.py [2] and while zuul_swift_upload provided the correct URL where the artifacts were supposed to be stored, it returned a 404. The uploaded files were really located at the same location with the ZUUL_UUID.

[1]: https://github.com/openstack-infra/project-config/blob/3295b44be291c3bab430efd4c5e5fcdedd82f49c/jenkins/scripts/zuul_swift_upload.py#L467
[2]: https://review.rdoproject.org/r/#/c/1652

Patch-set: 1