Update patch set 1
Patch Set 1: Code-Review-1

Ah, yeah it doesn't look like this is using the local cache -- however, I think the solution to this should be different.

The idea was that the launch server would maintain an up to date local cache of all the git repos.  Then, when a job arrives, it creates ephemeral git repos in the JobDir as needed.  It *should* clone them from the local cache (ideally with hard links), but it looks like that's the part I forgot to implement.  :)  Then it should perform the speculative merge on that JobDir git repo, and that's what gets pushed to the worker.

The reason for this approach is that we expect a launch server to run many jobs at once, and we don't necessarily want to serialize the speculative merging for all of them (in addition to the actual time spent merging, there's also the time that will be spent pushing the contents onto the worker).  This approach also handily deals with the cleanup problem -- the ephemeral refs are only in the JobDir git repos, so they are cleaned up automatically.

Patch-set: 1
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1