Update patch set 36
Patch Set 36: Code-Review-1

(7 comments)

We will be in the 1 million rows range within 2 months, so I think we need to be very careful with the schema.  We should learn from subunit2sql (which is storing some of the same data at the same scale) and make sure that SpamapS and mtreinish review at least the final version before we land it.

Patch-set: 36
Label: Code-Review=-1