Remove scheduler.getProject
Note that this removes the vestigal foreign project code.  That
feature has not currently been reworked into v3 and may need to
be in the future.  It the mean time, in order to better support
recording job provenance (ie, what repo a job was defined in)
we need to ensure that Projects are globally unique entities.
With this patch, Projects are only ever created by source objects.

Change-Id: I094cb6f82e06317ee08862757cf306e83576c47d