Update patch set 5
Patch Set 5: Code-Review-1

(7 comments)

Am I missing something about the project-templates, or are they not documented in the zuulv3 spec? If they are missing could you please add a few words in the spec about them to help me parse some of this?

I think I understand it, but I'm not sure what the endgame is. The -1 is for the inline comments.

Patch-set: 5
Label: Code-Review=-1