Do not import paramiko before daemonization
Paramiko relies on Crypto.Random() which opens a file descriptor on
/dev/urandom.  When zuul-server is daemonized, python-daemon closes all
file descriptor and read() attempt on it would thus lock.  That prevents
the Gerrit SSH connection from acquiring random number and deadlock
Zuul.

The root cause is the refactoring of Gerrit connetions. The ZuulApp
import zuul.lib.connetions which imports paramiko.  That happens before
the daemonization.

Move the import next to the method that relies on it.

A better fix will be to bump paramiko to 1.11.6 which relies on
os.urandom as has been done.  That has been done in Nodepool
https://review.openstack.org/#/c/269303/ but that is less trivial and
needs statsd to be bumped as well.

Change-Id: I4fb772f098d2927962e3f2df987ed17390a73c2e