Update patch set 5
Patch Set 5:

(4 comments)

Project templates are intended to continue to work more or less the same as in v2, so the spec says:

"""Templates are still supported. If a project lists a job that is defined in a template that is also applied to that project, the project-local specification of the job will modify that supplied by the template."""

And that's all it has to say on the subject.

Patch-set: 5