Clear up merger functions
Deobfuscate mergeChanges to only do merges (and not cat out files).
Return True or False on success or failure.

Use merge's getFiles to cat file (such as remote zuul confs).

This has the downside of not getting the files at the same time as
merge and therefore not being ready once the build_set is complete.
However, the mergers already have the commit so grabbing the files
should be fast.

Change-Id: Ia47ee2a57f9b36d4f84bbaca14d13314af7f12e5