Update patch set 3
Patch Set 3: Code-Review-1

(2 comments)

What we're saying here is that any static-node isn't allowed to be offline. This might be the correct solution, but we should consider if there is a case where we want a special job to all of a sudden take offline a static node. It could, for example, be helpful with debugging.

Patch-set: 3
Label: Code-Review=-1