(wip) Treat builds without a registered job as cancelled
This patch achieves a desired result of not having zuul post
NOT_REGISTERED comments to gerrit, but it is most likely not the correct
way to do this.

The NOT_REGISTERED comment is more useful for CI operator, less for
developers seeing the comments. The NOT_REGISTERED usually hints on CI
misconfiguration, or even just simple gearman restart, when nothing is
wrong with the CI.

There needs to be a way to either not comment in this case, or avoid the
underlying reason.

A possible solution would be instead of giving up and commenting with
the error, to actually register missing job with gearman, but not give
it any workers. This way job would be enqueued, and run as soon as valid
worker gets registered (or remain in queue forever).

Change-Id: Ie6d5ea35c6eeed465168f24921b04442df8f5744