Update patch set 5
Patch Set 5:

Joshua Hesketh wrote:

> I wonder if we'd get better coverage if we subclassed the ansible
> launcher to add in the features we need to introspect during testing
> rather than faking out the worker entirely.

That's what the AnsibleZuulTestCase does -- it runs the actual v3 ansible launcher, and does so with a subclassed gearman worker that lets us introspect.

I'm reinstating the fake worker because the ansible launcher is slower and will be cumbersome for working with jobs that we need to pause and later release in order to fully exercise some of the features of the scheduler.  But we will have both kinds of workers available to tests.

Patch-set: 5