Tests: Improve support in fake builds
This change makes two improvments to FakeBuild:

1) Restore support for examining commits present in a build

Since the merge for a commit is now performed on the launch server,
the previous method of examining the git repo state on the merger
no longer works.  Move the method that performs this function to
the FakeBuild and give the FakeBuild the context of the launch
server's JobDir.  Then it can inspect the repo states as created
in the JobDir on the launch server.  These are the repos that will
be pushed onto the test node.

The ZUUL_COMMIT env variable is no longer revelant, so remove that
from the check.

2) Restore support for releasing a held build.

Change-Id: I654a269d37c0bc323ed73afa68a73ddd558be7e2