Stub out aborting jobs in ansible launcher
The scheduler tests need to be able to abort jobs.  Even though
our Ansible launcher is not capable of aborting jobs yet, implement
the method so that we can override it in the RecordingLaunchServer
and exercise the scheduler.

Instead of using name+number to specify which build to abort, use
the UUID of the build.

Change-Id: I75525a2b48eb2761d599c039ba3084f09609dfbe