Add __repr__ to a few classes
When validating the layout, various objects are shown with the default
python representation which has a varying object id.  If one wants to
validate the difference of layout between two configuration, it ends up
with lot of noise such as:

 INFO:zuul.IndependentPipelineManager:  On success:
-INFO:zuul.IndependentPipelineManager:
[<zuul.reporter.gerrit.GerritReporter object at 0x4a04210>]
+INFO:zuul.IndependentPipelineManager:
[<zuul.reporter.gerrit.GerritReporter object at 0x357e210>]

To BaseConnection class:
* add __repr__ to show the concrete class name followed by the
  connection name.
* add a __str__ that is human/inclusion friendly using an URL like
  scheme: driver_name://connectio_name. Eg: gerrit://my_review_site

The base class for reporters, sources and triggers receive a __repr__
which include the connection friendly name.

Downstream bug:
https://phabricator.wikimedia.org/T143966

Change-Id: Ib7f1e1724e90160bbaefd581ed6b243728679892