Update patch set 6
Patch Set 6: Code-Review-1

After a bit of chatting in IRC, I agree that this is something that we should handle more gracefully.  However, this patch sets a value that will end up being applied to all of the git repos, and we can't be certain that the reset_branch will be valid for every one.

So instead, maybe we could, in the case that HEAD does not exist, just pick an arbitrary remote branch and reset to that.  Maybe just the first one in the remote branch list.  The branch shouldn't be important for this -- it's just trying to get the local repo in a known good state before the next checkout.

Patch-set: 6
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1