Update patch set 3
Patch Set 3:

Currently we have a number of jobs which encounter ssh connection issues.  Because the async module treats these as failures (not unreachable), it actually causes the job to fail, as opposed to most ssh connection issues which cause it to automatically relaunch.

About half of these jobs have a single ssh failure, fail, and then go on to sucessfully collect logs (these are the ones that really fail the job).  The other half have further ssh connection failures, but since they aren't in the async module, they are recorded as 'unreachable'.  *That* triggers the auto-relaunch behavior in Zuul.

While I'm okay with this change in general, I'm concerned that if we land it now (while still using async, and with async exhibiting this behavior), we will increase the failure rate (by not having an opportunity to fail a second time which causes the job to re-launch -- it's all backwards, I know).  So I'd like to hold off until we deal with that.

Patch-set: 3
Label: Verified=0