Add async action plugin to override upstream async
This action plugin is different than the upstream async action plugin in
two ways. It doesn't use the async_runner/async_status dance, but
instead just uses a forked process, a while loop and a timeout on the
calling host. Second, it's aware of our timeout callback plugin, so it
can handle the cumulative timeout without the need for it to be in the
playbook itself.

Also, made the cumulative timeout round up instead of down, largely so
that in testing it was easier to track that something was happening.

The normal action plugin operates exactly the same, so that we can still
timeout tasks from playbooks that don't declare themselves async.

Change-Id: If500ccc32e554cf4079f23fa3f215f5a6665a2a3