Extract makeBuilderTask and writeMainPlaybook into functions
makeBuilderTask is now not dependent on any runtime qualities, since
they are emitted into the vars.yaml file. Take it out of the class
and into standalone functions so that they can be used in a conversion
tool. Also, refactor the loop that creates the main playbook into
a standalone function for the same reason.

Change-Id: Iab5519441ff952225205bcc2f33364346f28e649