Update patch set 39
Patch Set 39:

(1 comment)

Okay so I agree having buildsets makes much more sense from zuul's point of view and we can figure out the os-health side of it later.

The reason for the metadata table was to be able to add in more fields over time (that's why there was only 2 to begin with). However we can always revisit that part of the schema with alembic migrations later. So I'll iterate on this with the proposed schema which lgtm.

Patch-set: 39
Label: Verified=0