Always query Gerrit if change was merged (Gerrit 2.13 fix)
Sometimes, a change is submitted before Zuul tries to do it, e.g.
manually in Gerrit during verification. In that case Gerrit, at least
in 2.13.1, reports "error: fatal: change is merged". Zuul then assumes
the change did not merge.

Always query Gerrit, independent of if the report worked or not, to see
if the change has been submitted, even if it was not Zuul.

Before this change, Zuul would repeatedly schedule gating of the merged
change in an infinite loop.

Signed-off-by: Fredrik Medley <fredrik.medley@autoliv.com>
Signed-off-by: Fredrik Medley <fredrik.medley@gmail.com>
Change-Id: I344d4360720191a2bba7d2fce0bd87cc36c1179b