Update patch set 1
Patch Set 1:

(If I understand it correctly, it's been a while since I looked). The ssh-wrapper is created by the merger server, but the key is private key is configured in zuul.conf. The hostkey for the merger server should be known by the puppet module and can be added to the known hosts of the mergers/scheduler etc.

Patch-set: 1