Ansible launcher: Fix afs publisher root detection
The find command that collected the marker files is expected
to print paths with a leading '/' (see later commands which
grep for '^/') but this was omitted.  This would cause all jobs
which published to the root (whether they had any content in
the root directory or were simply only intended to publish to a
subdir of the root) to conflict with each other.

Also, correct a missing fully-qualified path.

Change-Id: I6030c2b101026ff8e72cf4043e1d1b4fbffc5dcb