Add tool for aborting jobs in jenkins
Add a script to the tools directory that can be used to abort jenkins
jobs. By default it will poll zuul every 30 seconds and look for changes
in check and gate pipelines with a failed job. If it finds a failed job,
it will abort other running jobs rather than let them run to completion.

This can be useful when running jobs that can take a long time to run
and slave resources are scarce.

The script can also optionally abort all running jobs, which can be
useful during planned outage windows.

Clearly this is only useful in legacy zuul setups that still use jenkins
masters for dispatching jobs.

Change-Id: Ia049f02299075f52a37bbbe9e799d3e4615865e7